refactor: discountRate 반환 타입을 BigDecimal로 유지, response는 toInt만 호출
- PointProductWithDiscount.discountRate(): BigDecimal (scale 0, HALF_UP 반올림 완료)
- Response는 product.discountRate().toInt()로 단순 변환만 수행
- 반올림 책임은 도메인에 유지, response는 타입 변환 역할만

+++ boot/ma-boot-web/src/main/kotlin/com/konkuk/ma/domain/point/api/response/PointProductResponse.kt
@@ -20,7 +20,7 @@ class PointProductResponse(
                 quantity = product.pointProduct.quantity,
                 price = product.pointProduct.price.toInt(),
                 discountedPrice = product.discountedPrice().toInt(),
-                discountRate = product.discountRate(),
+                discountRate = product.discountRate().toInt(),
                 discountType = product.discountType()?.name,
                 isDiscountActive = product.isDiscountActive(),
             )
